changing ui

+++ ProjetWebTrekSwap/src/Controller/ReclamationController.php
@@ -24,7 +24,7 @@ final class ReclamationController extends AbstractController {
 
     // Front office routes
     #[Route('/new', name: 'app_reclamation_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, ReclamationRepository $reclamationRepository): Response
     {
         $reclamation = new Reclamation();
         $form = $this->createForm(ReclamationType::class, $reclamation);
@@ -41,6 +41,7 @@ final class ReclamationController extends AbstractController {
         return $this->render('reclamation/new.html.twig', [
             'reclamation' => $reclamation,
             'form' => $form,
+            'reclamations' => $reclamationRepository->findAll(),
         ]);
     }
 
@@ -56,8 +57,14 @@ final class ReclamationController extends AbstractController {
 
     #[Route('/back/{id_rec}', name: 'app_reclamation_show', methods: ['GET'])]
     #[IsGranted('ROLE_ADMIN')]
-    public function show(Reclamation $reclamation): Response
+    public function show(ReclamationRepository $reclamationRepository, int $id_rec): Response
     {
+        $reclamation = $reclamationRepository->findOneBy(['idRec' => $id_rec]);
+        
+        if (!$reclamation) {
+            throw $this->createNotFoundException('La réclamation n\'existe pas');
+        }
+
         return $this->render('reclamation/show.html.twig', [
             'reclamation' => $reclamation,
         ]);
@@ -65,8 +72,14 @@ final class ReclamationController extends AbstractController {
 
     #[Route('/back/{id_rec}/edit', name: 'app_reclamation_edit', methods: ['GET', 'POST'])]
     #[IsGranted('ROLE_ADMIN')]
-    public function edit(Request $request, Reclamation $reclamation, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, ReclamationRepository $reclamationRepository, int $id_rec, EntityManagerInterface $entityManager): Response
     {
+        $reclamation = $reclamationRepository->findOneBy(['idRec' => $id_rec]);
+        
+        if (!$reclamation) {
+            throw $this->createNotFoundException('La réclamation n\'existe pas');
+        }
+
         $form = $this->createForm(ReclamationType::class, $reclamation);
         $form->handleRequest($request);
 
@@ -85,8 +98,14 @@ final class ReclamationController extends AbstractController {
 
     #[Route('/back/{id_rec}', name: 'app_reclamation_delete', methods: ['POST'])]
     #[IsGranted('ROLE_ADMIN')]
-    public function delete(Request $request, Reclamation $reclamation, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, ReclamationRepository $reclamationRepository, int $id_rec, EntityManagerInterface $entityManager): Response
     {
+        $reclamation = $reclamationRepository->findOneBy(['idRec' => $id_rec]);
+        
+        if (!$reclamation) {
+            throw $this->createNotFoundException('La réclamation n\'existe pas');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$reclamation->getIdRec(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($reclamation);
             $entityManager->flush();
